authorization implemented

+++ src/api/notes/handler.js
@@ -1,20 +1,30 @@
+const autoBind = require("auto-bind");
+
 class NotesHandler {
     constructor(service, validator) {
         this._service = service;
         this._validator = validator;
 
-        this.postNoteHandler = this.postNoteHandler.bind(this);
-        this.getNotesHandler = this.getNotesHandler.bind(this);
-        this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
-        this.putNoteByIdHandler = this.putNoteByIdHandler.bind(this);
-        this.deleteNoteByIdHandler = this.deleteNoteByIdHandler.bind(this);
+        autoBind(this);
+        // this.postNoteHandler = this.postNoteHandler.bind(this);
+        // this.getNotesHandler = this.getNotesHandler.bind(this);
+        // this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
+        // this.putNoteByIdHandler = this.putNoteByIdHandler.bind(this);
+        // this.deleteNoteByIdHandler = this.deleteNoteByIdHandler.bind(this);
     }
 
     async postNoteHandler(request, h) {
         this._validator.validateNotePayload(request.payload);
         const { title = "untitled", body, tags } = request.payload;
 
-        const noteId = await this._service.addNote({ title, body, tags });
+        const { id: credentialId } = request.auth.credentials;
+
+        const noteId = await this._service.addNote({
+            title,
+            body,
+            tags,
+            owner: credentialId,
+        });
 
         const response = h.response({
             status: "success",
@@ -27,31 +37,36 @@ class NotesHandler {
         return response;
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes();
+    async getNotesHandler(request, h) {
+        const { id: credentialId } = request.auth.credentials;
+
+        const notes = await this._service.getNotes(credentialId);
+
         return {
             status: "success",
-            data: {
-                notes,
-            },
+            data: { notes },
         };
     }
 
     async getNoteByIdHandler(request, h) {
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
+
+        await this._service.verifyNoteOwner(id, credentialId);
         const note = await this._service.getNoteById(id);
+
         return {
             status: "success",
-            data: {
-                note,
-            },
+            data: { note },
         };
     }
 
     async putNoteByIdHandler(request, h) {
         this._validator.validateNotePayload(request.payload);
         const { id } = request.params;
+        const { id: credentialId } = request.auth.credentials;
 
+        await this._service.verifyNoteOwner(id, credentialId);
         await this._service.editNoteById(id, request.payload);
 
         return {
@@ -62,7 +77,10 @@ class NotesHandler {
 
     async deleteNoteByIdHandler(request, h) {
         const { id } = request.params;
-        this._service.deleteNoteById(id);
+        const { id: credentialId } = request.auth.credentials;
+
+        await this._service.verifyNoteOwner(id, credentialId);
+        await this._service.deleteNoteById(id);
 
         return {
             status: "success",
